app/requirements.txt
cassandra-driver
more-itertools